Moved tractor beam logic into its own controller. Relies on the PlayerController to call it at the appropriate context however.
