+++ Assets/Scenes/Main.unity
@@ -4995,7 +4995,7 @@ MonoBehaviour:
   m_CorrespondingSourceObject: {fileID: 4477257796416552750, guid: 6cd717d2d8734544281e5313ec627f85, type: 3}
   m_PrefabInstance: {fileID: 4477257797740222008}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
+  m_GameObject: {fileID: 4477257797740222009}
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: afb911a6128e9564d8c7ae860fc9abd9, type: 3}
@@ -5271,6 +5271,23 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 6cd717d2d8734544281e5313ec627f85, type: 3}
+--- !u!1 &4477257797740222009 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 4477257796416552756, guid: 6cd717d2d8734544281e5313ec627f85, type: 3}
+  m_PrefabInstance: {fileID: 4477257797740222008}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &4477257797740222010
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4477257797740222009}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e5066621c47170444b86adfedf547bdc, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1001 &4643123014844411523
 PrefabInstance:
   m_ObjectHideFlags: 0
