+++ Assets/Scripts/Game/Controllers/TractorBeamController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e5066621c47170444b86adfedf547bdc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
