+++ UserSettings/EditorUserSettings.asset
@@ -27,10 +27,10 @@ EditorUserSettings:
       value: 22424703114646680e0b0227036c6f1504041139382d2824633c133af6f9
       flags: 0
     RecentlyUsedScenePath-7:
-      value: 22424703114646680e0b0227036c72111f19152f223d68252320092a
+      value: 22424703114646680e0b0227036c6d0518240c2b3e3c68252320092a
       flags: 0
     RecentlyUsedScenePath-8:
-      value: 22424703114646680e0b0227036c6d0518240c2b3e3c68252320092a
+      value: 22424703114646680e0b0227036c72111f19152f223d68252320092a
       flags: 0
     RecentlyUsedScenePath-9:
       value: 22424703114646680e0b0227036c72111f19563f22213229
